Cleanup shared's reexports
