+++ egui_commonmark/examples/macro.rs
@@ -1,12 +1,11 @@
 //! Make sure to run this example from the repo directory and not the example
 //! directory. To see all the features in full effect, run this example with
-//! `cargo r --features better_syntax_highlighting,svg,fetch`
+//! `cargo r --example macro --features macro,better_syntax_highlighting`
 //! Add `light` or `dark` to the end of the command to specify theme. Default
-//! is light. `cargo r --features better_syntax_highlighting,svg,fetch -- dark`
+//! is light. `cargo r --example macro --features macro,better_syntax_highlighting -- dark`
 
 use eframe::egui;
 use egui_commonmark::*;
-use egui_commonmark_macros::*;
 
 struct App {
     cache: CommonMarkCache,
