+++ egui_commonmark/src/lib.rs
@@ -33,7 +33,7 @@ use egui::{self, Id};
 mod parsers;
 
 pub use egui_commonmark_shared::alerts::{Alert, AlertBundle};
-pub use egui_commonmark_shared::CommonMarkCache;
+pub use egui_commonmark_shared::misc::CommonMarkCache;
 
 #[cfg(all(feature = "comrak", feature = "pulldown_cmark"))]
 compile_error!("Cannot have multiple different parsing backends enabled at the same time");
@@ -42,7 +42,13 @@ compile_error!("Cannot have multiple different parsing backends enabled at the s
 compile_error!("Either the pulldown_cmark or comrak backend must be enabled");
 
 #[cfg(feature = "macro")]
+pub use egui_commonmark_macros::*;
+
+#[cfg(feature = "macro")]
+// Do not rely on this directly!
+#[doc(hidden)]
 pub use egui_commonmark_shared;
+
 use egui_commonmark_shared::*;
 
 #[derive(Debug)]
