+++ egui_commonmark_shared/src/lib.rs
@@ -1,548 +1,21 @@
 //! Shared code for egui_commonmark and egui_commonmark_macro. Breaking changes will happen and
 //! should not be relied upon directly
 
-#[cfg(feature = "better_syntax_highlighting")]
-use syntect::{
-    easy::HighlightLines,
-    highlighting::{Theme, ThemeSet},
-    parsing::{SyntaxDefinition, SyntaxSet},
-    util::LinesWithEndings,
-};
-
-use egui::{text::LayoutJob, RichText, TextStyle, Ui};
-use std::collections::HashMap;
-
+#[doc(hidden)]
 pub mod alerts;
+#[doc(hidden)]
 pub mod elements;
+#[doc(hidden)]
+pub mod misc;
 #[cfg(feature = "pulldown-cmark")]
+#[doc(hidden)]
 pub mod pulldown;
 
 // For ease of use in proc macros
-pub use alerts::*;
-pub use elements::*;
-
-#[cfg(feature = "better_syntax_highlighting")]
-const DEFAULT_THEME_LIGHT: &str = "base16-ocean.light";
-#[cfg(feature = "better_syntax_highlighting")]
-const DEFAULT_THEME_DARK: &str = "base16-ocean.dark";
-
-#[derive(Debug)]
-pub struct CommonMarkOptions {
-    pub indentation_spaces: usize,
-    pub max_image_width: Option<usize>,
-    pub show_alt_text_on_hover: bool,
-    pub default_width: Option<usize>,
-    #[cfg(feature = "better_syntax_highlighting")]
-    pub theme_light: String,
-    #[cfg(feature = "better_syntax_highlighting")]
-    pub theme_dark: String,
-    pub use_explicit_uri_scheme: bool,
-    pub default_implicit_uri_scheme: String,
-    pub alerts: AlertBundle,
-    /// Whether to present a mutable ui for things like checkboxes
-    pub mutable: bool,
-}
-
-impl Default for CommonMarkOptions {
-    fn default() -> Self {
-        Self {
-            indentation_spaces: 4,
-            max_image_width: None,
-            show_alt_text_on_hover: true,
-            default_width: None,
-            #[cfg(feature = "better_syntax_highlighting")]
-            theme_light: DEFAULT_THEME_LIGHT.to_owned(),
-            #[cfg(feature = "better_syntax_highlighting")]
-            theme_dark: DEFAULT_THEME_DARK.to_owned(),
-            use_explicit_uri_scheme: false,
-            default_implicit_uri_scheme: "file://".to_owned(),
-            alerts: AlertBundle::gfm(),
-            mutable: false,
-        }
-    }
-}
-
-impl CommonMarkOptions {
-    #[cfg(feature = "better_syntax_highlighting")]
-    pub fn curr_theme(&self, ui: &Ui) -> &str {
-        if ui.style().visuals.dark_mode {
-            &self.theme_dark
-        } else {
-            &self.theme_light
-        }
-    }
-
-    pub fn max_width(&self, ui: &Ui) -> f32 {
-        let max_image_width = self.max_image_width.unwrap_or(0) as f32;
-        let available_width = ui.available_width();
-
-        let max_width = max_image_width.max(available_width);
-        if let Some(default_width) = self.default_width {
-            if default_width as f32 > max_width {
-                default_width as f32
-            } else {
-                max_width
-            }
-        } else {
-            max_width
-        }
-    }
-}
-
-#[derive(Default, Clone)]
-pub struct Style {
-    pub heading: Option<u8>,
-    pub strong: bool,
-    pub emphasis: bool,
-    pub strikethrough: bool,
-    pub quote: bool,
-    pub code: bool,
-}
-
-impl Style {
-    pub fn to_richtext(&self, ui: &Ui, text: &str) -> RichText {
-        let mut text = RichText::new(text);
-
-        if let Some(level) = self.heading {
-            let max_height = ui
-                .style()
-                .text_styles
-                .get(&TextStyle::Heading)
-                .map_or(32.0, |d| d.size);
-            let min_height = ui
-                .style()
-                .text_styles
-                .get(&TextStyle::Body)
-                .map_or(14.0, |d| d.size);
-            let diff = max_height - min_height;
-
-            match level {
-                0 => {
-                    text = text.strong().heading();
-                }
-                1 => {
-                    let size = min_height + diff * 0.835;
-                    text = text.strong().size(size);
-                }
-                2 => {
-                    let size = min_height + diff * 0.668;
-                    text = text.strong().size(size);
-                }
-                3 => {
-                    let size = min_height + diff * 0.501;
-                    text = text.strong().size(size);
-                }
-                4 => {
-                    let size = min_height + diff * 0.334;
-                    text = text.size(size);
-                }
-                // We only support 6 levels
-                5.. => {
-                    let size = min_height + diff * 0.167;
-                    text = text.size(size);
-                }
-            }
-        }
-
-        if self.quote {
-            text = text.weak();
-        }
-
-        if self.strong {
-            text = text.strong();
-        }
-
-        if self.emphasis {
-            // FIXME: Might want to add some space between the next text
-            text = text.italics();
-        }
-
-        if self.strikethrough {
-            text = text.strikethrough();
-        }
-
-        if self.code {
-            text = text.code();
-        }
-
-        text
-    }
-}
-
-#[derive(Default)]
-pub struct Link {
-    pub destination: String,
-    pub text: Vec<RichText>,
-}
-
-impl Link {
-    pub fn end(self, ui: &mut Ui, cache: &mut CommonMarkCache) {
-        let Self { destination, text } = self;
-
-        let mut layout_job = LayoutJob::default();
-        for t in text {
-            t.append_to(
-                &mut layout_job,
-                ui.style(),
-                egui::FontSelection::Default,
-                egui::Align::LEFT,
-            );
-        }
-        if cache.link_hooks().contains_key(&destination) {
-            let ui_link = ui.link(layout_job);
-            if ui_link.clicked() || ui_link.middle_clicked() {
-                cache.link_hooks_mut().insert(destination, true);
-            }
-        } else {
-            ui.hyperlink_to(layout_job, destination);
-        }
-    }
-}
-
-pub struct Image {
-    pub uri: String,
-    pub alt_text: Vec<RichText>,
-}
-
-impl Image {
-    // FIXME: string conversion
-    pub fn new(uri: &str, options: &CommonMarkOptions) -> Self {
-        let has_scheme = uri.contains("://");
-        let uri = if options.use_explicit_uri_scheme || has_scheme {
-            uri.to_string()
-        } else {
-            // Assume file scheme
-            format!("{}{uri}", options.default_implicit_uri_scheme)
-        };
-
-        Self {
-            uri,
-            alt_text: Vec::new(),
-        }
-    }
-
-    pub fn end(self, ui: &mut Ui, options: &CommonMarkOptions) {
-        let response = ui.add(
-            egui::Image::from_uri(&self.uri)
-                .fit_to_original_size(1.0)
-                .max_width(options.max_width(ui)),
-        );
-
-        if !self.alt_text.is_empty() && options.show_alt_text_on_hover {
-            response.on_hover_ui_at_pointer(|ui| {
-                for alt in self.alt_text {
-                    ui.label(alt);
-                }
-            });
-        }
-    }
-}
-
-pub struct FencedCodeBlock {
-    pub lang: String,
-    pub content: String,
-}
-
-impl FencedCodeBlock {
-    pub fn end(
-        &self,
-        ui: &mut Ui,
-        cache: &mut CommonMarkCache,
-        options: &CommonMarkOptions,
-        max_width: f32,
-    ) {
-        ui.scope(|ui| {
-            Self::pre_syntax_highlighting(cache, options, ui);
-
-            let mut layout = |ui: &Ui, string: &str, wrap_width: f32| {
-                let mut job = self.syntax_highlighting(cache, options, &self.lang, ui, string);
-                job.wrap.max_width = wrap_width;
-                ui.fonts(|f| f.layout_job(job))
-            };
-
-            elements::code_block(ui, max_width, &self.content, &mut layout);
-        });
-    }
-}
-
-#[cfg(not(feature = "better_syntax_highlighting"))]
-impl FencedCodeBlock {
-    fn pre_syntax_highlighting(
-        _cache: &mut CommonMarkCache,
-        _options: &CommonMarkOptions,
-        ui: &mut Ui,
-    ) {
-        ui.style_mut().visuals.extreme_bg_color = ui.visuals().extreme_bg_color;
-    }
-
-    fn syntax_highlighting(
-        &self,
-        _cache: &mut CommonMarkCache,
-        _options: &CommonMarkOptions,
-        extension: &str,
-        ui: &Ui,
-        text: &str,
-    ) -> egui::text::LayoutJob {
-        crate::plain_highlighting(ui, text, extension)
-    }
-}
-
-#[cfg(feature = "better_syntax_highlighting")]
-impl FencedCodeBlock {
-    fn pre_syntax_highlighting(
-        cache: &mut CommonMarkCache,
-        options: &CommonMarkOptions,
-        ui: &mut Ui,
-    ) {
-        let curr_theme = cache.curr_theme(ui, options);
-        let style = ui.style_mut();
-
-        style.visuals.extreme_bg_color = curr_theme
-            .settings
-            .background
-            .map(crate::syntect_color_to_egui)
-            .unwrap_or(style.visuals.extreme_bg_color);
-
-        if let Some(color) = curr_theme.settings.selection_foreground {
-            style.visuals.selection.bg_fill = crate::syntect_color_to_egui(color);
-        }
-    }
-
-    fn syntax_highlighting(
-        &self,
-        cache: &CommonMarkCache,
-        options: &CommonMarkOptions,
-        extension: &str,
-        ui: &Ui,
-        text: &str,
-    ) -> egui::text::LayoutJob {
-        if let Some(syntax) = cache.ps.find_syntax_by_extension(extension) {
-            let mut job = egui::text::LayoutJob::default();
-            let mut h = HighlightLines::new(syntax, cache.curr_theme(ui, options));
-
-            for line in LinesWithEndings::from(text) {
-                let ranges = h.highlight_line(line, &cache.ps).unwrap();
-                for v in ranges {
-                    let front = v.0.foreground;
-                    job.append(
-                        v.1,
-                        0.0,
-                        egui::TextFormat::simple(
-                            TextStyle::Monospace.resolve(ui.style()),
-                            crate::syntect_color_to_egui(front),
-                        ),
-                    );
-                }
-            }
-
-            job
-        } else {
-            crate::plain_highlighting(ui, text, extension)
-        }
-    }
-}
-
-fn plain_highlighting(ui: &Ui, text: &str, extension: &str) -> egui::text::LayoutJob {
-    egui_extras::syntax_highlighting::highlight(
-        ui.ctx(),
-        &egui_extras::syntax_highlighting::CodeTheme::from_style(ui.style()),
-        text,
-        extension,
-    )
-}
-
-#[cfg(feature = "better_syntax_highlighting")]
-fn syntect_color_to_egui(color: syntect::highlighting::Color) -> egui::Color32 {
-    egui::Color32::from_rgb(color.r, color.g, color.b)
-}
-
-#[cfg(feature = "better_syntax_highlighting")]
-fn default_theme(ui: &Ui) -> &str {
-    if ui.style().visuals.dark_mode {
-        DEFAULT_THEME_DARK
-    } else {
-        DEFAULT_THEME_LIGHT
-    }
-}
-
-#[cfg(feature = "pulldown-cmark")]
-use pulldown::ScrollableCache;
-
-/// A cache used for storing content such as images.
-#[derive(Debug)]
-pub struct CommonMarkCache {
-    // Everything stored in `CommonMarkCache` must take into account that
-    // the cache is for multiple `CommonMarkviewer`s with different source_ids.
-    #[cfg(feature = "better_syntax_highlighting")]
-    ps: SyntaxSet,
-
-    #[cfg(feature = "better_syntax_highlighting")]
-    ts: ThemeSet,
-
-    link_hooks: HashMap<String, bool>,
-
-    #[cfg(feature = "pulldown-cmark")]
-    scroll: HashMap<egui::Id, ScrollableCache>,
-    pub(self) has_installed_loaders: bool,
-}
-
-#[allow(clippy::derivable_impls)]
-impl Default for CommonMarkCache {
-    fn default() -> Self {
-        Self {
-            #[cfg(feature = "better_syntax_highlighting")]
-            ps: SyntaxSet::load_defaults_newlines(),
-            #[cfg(feature = "better_syntax_highlighting")]
-            ts: ThemeSet::load_defaults(),
-            link_hooks: HashMap::new(),
-            #[cfg(feature = "pulldown-cmark")]
-            scroll: Default::default(),
-            has_installed_loaders: false,
-        }
-    }
-}
-
-impl CommonMarkCache {
-    #[cfg(feature = "better_syntax_highlighting")]
-    pub fn add_syntax_from_folder(&mut self, path: &str) {
-        let mut builder = self.ps.clone().into_builder();
-        let _ = builder.add_from_folder(path, true);
-        self.ps = builder.build();
-    }
-
-    #[cfg(feature = "better_syntax_highlighting")]
-    pub fn add_syntax_from_str(&mut self, s: &str, fallback_name: Option<&str>) {
-        let mut builder = self.ps.clone().into_builder();
-        let _ = SyntaxDefinition::load_from_str(s, true, fallback_name).map(|d| builder.add(d));
-        self.ps = builder.build();
-    }
-
-    #[cfg(feature = "better_syntax_highlighting")]
-    /// Add more color themes for code blocks(.tmTheme files). Set the color theme with
-    /// [`syntax_theme_dark`](CommonMarkViewer::syntax_theme_dark) and
-    /// [`syntax_theme_light`](CommonMarkViewer::syntax_theme_light)
-    pub fn add_syntax_themes_from_folder(
-        &mut self,
-        path: impl AsRef<std::path::Path>,
-    ) -> Result<(), syntect::LoadingError> {
-        self.ts.add_from_folder(path)
-    }
-
-    #[cfg(feature = "better_syntax_highlighting")]
-    /// Add color theme for code blocks(.tmTheme files). Set the color theme with
-    /// [`syntax_theme_dark`](CommonMarkViewer::syntax_theme_dark) and
-    /// [`syntax_theme_light`](CommonMarkViewer::syntax_theme_light)
-    pub fn add_syntax_theme_from_bytes(
-        &mut self,
-        name: impl Into<String>,
-        bytes: &[u8],
-    ) -> Result<(), syntect::LoadingError> {
-        let mut cursor = std::io::Cursor::new(bytes);
-        self.ts
-            .themes
-            .insert(name.into(), ThemeSet::load_from_reader(&mut cursor)?);
-        Ok(())
-    }
-
-    /// Clear the cache for all scrollable elements
-    #[cfg(feature = "pulldown-cmark")]
-    pub fn clear_scrollable(&mut self) {
-        self.scroll.clear();
-    }
-
-    /// Clear the cache for a specific scrollable viewer. Returns false if the
-    /// id was not in the cache.
-    #[cfg(feature = "pulldown-cmark")]
-    pub fn clear_scrollable_with_id(&mut self, source_id: impl std::hash::Hash) -> bool {
-        self.scroll.remove(&egui::Id::new(source_id)).is_some()
-    }
-
-    /// If the user clicks on a link in the markdown render that has `name` as a link. The hook
-    /// specified with this method will be set to true. It's status can be acquired
-    /// with [`get_link_hook`](Self::get_link_hook). Be aware that all hook state is reset once
-    /// [`CommonMarkViewer::show`] gets called
-    ///
-    /// # Why use link hooks
-    ///
-    /// egui provides a method for checking links afterwards so why use this instead?
-    ///
-    /// ```rust
-    /// # use egui::__run_test_ctx;
-    /// # __run_test_ctx(|ctx| {
-    /// ctx.output_mut(|o| o.open_url.is_some());
-    /// # });
-    /// ```
-    ///
-    /// The main difference is that link hooks allows egui_commonmark to check for link hooks
-    /// while rendering. Normally when hovering over a link, egui_commonmark will display the full
-    /// url. With link hooks this feature is disabled, but to do that all hooks must be known.
-    pub fn add_link_hook<S: Into<String>>(&mut self, name: S) {
-        self.link_hooks.insert(name.into(), false);
-    }
-
-    /// Returns None if the link hook could not be found. Returns the last known status of the
-    /// hook otherwise.
-    pub fn remove_link_hook(&mut self, name: &str) -> Option<bool> {
-        self.link_hooks.remove(name)
-    }
-
-    /// Get status of link. Returns true if it was clicked
-    pub fn get_link_hook(&self, name: &str) -> Option<bool> {
-        self.link_hooks.get(name).copied()
-    }
-
-    /// Remove all link hooks
-    pub fn link_hooks_clear(&mut self) {
-        self.link_hooks.clear();
-    }
-
-    /// All link hooks
-    pub fn link_hooks(&self) -> &HashMap<String, bool> {
-        &self.link_hooks
-    }
-
-    /// Raw access to link hooks
-    pub fn link_hooks_mut(&mut self) -> &mut HashMap<String, bool> {
-        &mut self.link_hooks
-    }
-
-    /// Set all link hooks to false
-    fn deactivate_link_hooks(&mut self) {
-        for v in self.link_hooks.values_mut() {
-            *v = false;
-        }
-    }
-
-    #[cfg(feature = "better_syntax_highlighting")]
-    fn curr_theme(&self, ui: &Ui, options: &CommonMarkOptions) -> &Theme {
-        self.ts
-            .themes
-            .get(options.curr_theme(ui))
-            // Since we have called load_defaults, the default theme *should* always be available..
-            .unwrap_or_else(|| &self.ts.themes[default_theme(ui)])
-    }
-}
-
-#[cfg(feature = "pulldown-cmark")]
-pub fn scroll_cache<'a>(cache: &'a mut CommonMarkCache, id: &egui::Id) -> &'a mut ScrollableCache {
-    if !cache.scroll.contains_key(id) {
-        cache.scroll.insert(*id, Default::default());
-    }
-    cache.scroll.get_mut(id).unwrap()
-}
-
-/// Should be called before any rendering
-pub fn prepare_show(cache: &mut CommonMarkCache, ctx: &egui::Context) {
-    if !cache.has_installed_loaders {
-        // Even though the install function can be called multiple times, its not the cheapest
-        // so we ensure that we only call it once.
-        // This could be done at the creation of the cache, however it is better to keep the
-        // cache free from egui's Ui and Context types as this allows it to be created before
-        // any egui instances. It also keeps the API similar to before the introduction of the
-        // image loaders.
-        egui_extras::install_image_loaders(ctx);
-        cache.has_installed_loaders = true;
-    }
-
-    cache.deactivate_link_hooks();
-}
+#[doc(hidden)]
+pub use {
+    alerts::{alert_ui, Alert, AlertBundle},
+    // Pretty much every single element in this module is used by the proc macros
+    elements::*,
+    misc::{prepare_show, CommonMarkCache, CommonMarkOptions, FencedCodeBlock, Image, Link},
+};
